soup-auth-negotiate: Introduce a blacklist of URIs where the GSSAPI authentication won't happen
Introduce a blacklist of URIs where the GSSAPI authentication won't happen and
the user will fallback to provide the credentials. The list will be initialized
from the SOUP_GSSAPI_BLACKLISTED_URIS environmental variable. As a part of this
change the SOUP_AUTH_TRUSTED_URIS was renamed to SOUP_GSSAPI_TRUSTED_URIS to
express that it belongs to the GSSAPI configuration.