websocket-connection: avoid sending data if we are closing the connection
We cannot trust the client to be nice and stop sending data when
we are closing. If we receive a ping when closing the connection
we need to avoid trying to send data after that.

Change the code to be more robust and to not try to write in the socket
if we are in the closed state.